+++ src/main/java/ru/practicum/shareit/handler/CustomExceptionHandler.java
@@ -1,6 +1,7 @@
-package ru.practicum.shareit.booking.controller;
+package ru.practicum.shareit.handler;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -10,47 +11,68 @@ import ru.practicum.shareit.exception.*;
 import javax.validation.ValidationException;
 
 @Slf4j
-@RestControllerAdvice(assignableTypes = BookingController.class)
-public class BookingExceptionHandler {
+@RestControllerAdvice
+public class CustomExceptionHandler {
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleBookingParameter(final ParameterException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleBookingNotValid(final ValidationException e) {
+    public ErrorResponse handleParameter(final ParameterException e) {
         log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleBookingNotAvailable(final NotAvailableException e) {
+    public ErrorResponse handleNotValid(final ValidationException e) {
+        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handleNotAvailable(final NotAvailableException e) {
+        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handleNotOwnOrComplete(final NotOwnOrCompleteThisBookingException e) {
         log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleBookingNotFound(final NotFoundException e) {
+    public ErrorResponse handleNotOwnItemOrBooking(final NotOwnItemOrNotOwnBookingException e) {
         log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleBookingNotOwn(final NotOwnException e) {
+    public ErrorResponse handleNotOwned(final NotOwnException e) {
         log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleBookingNotOwnItemOrBooking(final NotOwnItemOrNotOwnBookingException e) {
+    public ErrorResponse handleNotFound(final NotFoundException e) {
+        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.CONFLICT)
+    public ErrorResponse handleEmailAlreadyExist(final AlreadyExistException e) {
+        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.CONFLICT)
+    public ErrorResponse handleNotUniq(final DataIntegrityViolationException e) {
         log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
         return new ErrorResponse(e.getMessage());
     }
