--- src/main/java/ru/practicum/shareit/item/controller/ItemExceptionHandler.java
@@ -1,57 +1,0 @@
-package ru.practicum.shareit.item.controller;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import ru.practicum.shareit.exception.*;
-
-import javax.validation.ValidationException;
-
-@Slf4j
-@RestControllerAdvice(assignableTypes = ItemController.class)
-public class ItemExceptionHandler {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleItemParameter(final ParameterException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleItemNotValid(final ValidationException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleItemNotAvailable(final NotAvailableException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleItemNotOwnOrComplete(final NotOwnOrCompleteThisBookingException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleItemNotFound(final NotFoundException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleItemNotOwned(final NotOwnException e) {
-        log.error("При обработке запроса возникла ошибка: {}.", e.getMessage());
-        return new ErrorResponse(e.getMessage());
-    }
-}
